+++ src/main/java/com/swyp3/skin/domain/user/service/UserService.java
@@ -56,8 +56,6 @@ public class UserService {
                 });
     }
 
-
-
     public MyPageResponse getMyPageInfo(Long userId) {
 
         UserOauth userOauth = userOauthRepository.findByUser_id(userId)
